requirements.txt
pytaglib==1.0.0
